Desenvolvimento de feature para exceções

+++ src/main/java/br/com/caelum/carangobom/impl/MarcaServiceImpl.java
@@ -7,17 +7,15 @@ import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
-import org.springframework.web.bind.MethodArgumentNotValidException;
 
 import br.com.caelum.carangobom.domain.Marca;
 import br.com.caelum.carangobom.domain.Veiculo;
+import br.com.caelum.carangobom.exception.MarcaAssociadaException;
 import br.com.caelum.carangobom.form.DashboardForm;
 import br.com.caelum.carangobom.form.MarcaForm;
 import br.com.caelum.carangobom.repository.MarcaRepository;
 import br.com.caelum.carangobom.repository.VeiculoRepository;
 import br.com.caelum.carangobom.service.MarcaService;
-import br.com.caelum.carangobom.validacao.ErroDeParametroOutputDto;
-import br.com.caelum.carangobom.validacao.ListaDeErrosOutputDto;
 
 
 @Service
@@ -33,10 +31,14 @@ public class MarcaServiceImpl implements MarcaService {
 	private MarcaForm marcaForm;
 
     @Override
-    public Marca removeBrand(Long id) {
+    public Marca removeBrand(Long id) throws MarcaAssociadaException {
     	Optional<Marca> m1 = marcaRepository.findById(id);
         if (m1.isPresent()) {
             Marca m2 = m1.get();
+            List<Veiculo> listVeiculos = veiculoRepository.findAllByMarcaId(m2.getId());
+            if(!listVeiculos.isEmpty() ) {
+            	throw new MarcaAssociadaException("Marca '" + m2.getNome() + "' está associada a veículos e não pode ser removida");
+            }
             marcaRepository.delete(m2);
             return m2;
         } else {
@@ -81,23 +83,6 @@ public class MarcaServiceImpl implements MarcaService {
             return null;
         }
 	}
-	
-	
-	@Override
-	public ListaDeErrosOutputDto validacao(MethodArgumentNotValidException excecao) {
-		List<ErroDeParametroOutputDto> l = new ArrayList<>();
-        excecao.getBindingResult().getFieldErrors().forEach(e -> {
-            ErroDeParametroOutputDto d = new ErroDeParametroOutputDto();
-            d.setParametro(e.getField());
-            d.setMensagem(e.getDefaultMessage());
-            l.add(d);
-        });
-        ListaDeErrosOutputDto l2 = new ListaDeErrosOutputDto();
-        l2.setErros(l);
-        return l2;
-		
-	}
-
 
 	@Override
 	public List<DashboardForm> findDashboard() {
@@ -112,11 +97,10 @@ public class MarcaServiceImpl implements MarcaService {
 	}
 	
 	private DashboardForm createMarcaDashboard(Marca marca, DashboardForm dashboard) {
-		Optional<List<Veiculo>> findVeiculosByMarcaId = veiculoRepository.findVeiculosByMarcaId(marca.getId().intValue());
-		if(findVeiculosByMarcaId.isPresent()) {
-			List<Veiculo> listVeiculo = findVeiculosByMarcaId.get();
-			dashboard.setQtdeVeiculos(extractQtdeVeiculoMarca(listVeiculo));
-			dashboard.setValorTotalVeiculos(extractValorVeiculosMarca(listVeiculo));
+		List<Veiculo> findVeiculosByMarcaId = veiculoRepository.findAllByMarcaIdAndIsVendido(marca.getId(), false);
+		if(findVeiculosByMarcaId != null && !findVeiculosByMarcaId.isEmpty()) {
+			dashboard.setQtdeVeiculos(extractQtdeVeiculoMarca(findVeiculosByMarcaId));
+			dashboard.setValorTotalVeiculos(extractValorVeiculosMarca(findVeiculosByMarcaId));
 		}
 		return dashboard;
 	}
